+++ vortex-array/src/array/varbin/compute/mod.rs
@@ -3,19 +3,14 @@ use vortex_scalar::Scalar;
 
 use crate::array::varbin::{varbin_scalar, VarBinArray};
 use crate::compute::unary::ScalarAtFn;
-use crate::compute::{ArrayCompute, MaybeCompareFn, Operator, SliceFn, TakeFn};
-use crate::{Array, ArrayDType};
+use crate::compute::{ArrayCompute, SliceFn, TakeFn};
+use crate::ArrayDType;
 
-mod compare;
 mod filter;
 mod slice;
 mod take;
 
 impl ArrayCompute for VarBinArray {
-    fn compare(&self, other: &Array, operator: Operator) -> Option<VortexResult<Array>> {
-        MaybeCompareFn::maybe_compare(self, other, operator)
-    }
-
     fn scalar_at(&self) -> Option<&dyn ScalarAtFn> {
         Some(self)
     }
